requirements.txt
numpy
scipy
pyvista
nibabel
pandas
pyyaml
pillow
typing_extensions
graphviz
imageio
imageio-ffmpeg